% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertColumnBreakAddin.R
\name{insertColumnBreakAddin}
\alias{insertColumnBreakAddin}
\title{Insert Rmarkdown column break}
\usage{
insertColumnBreakAddin()
}
\description{
Insert Rmarkdown code to end a column div and start another
}
