% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertColumnsAddin.R
\name{insertColumnsAddin}
\alias{insertColumnsAddin}
\title{Insert Rmarkdown divs for a two-column layout}
\usage{
insertColumnsAddin()
}
\description{
Insert Rmarkdown divs for a two-column layout into the current
document. Uses the pandoc `:::` ... `:::` syntax for divs,
with `{.columns}` and `{.column}` tags for column sets and individual
columns.
}
