% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDisplayAlignAddin.R
\name{insertDisplayAlignAddin}
\alias{insertDisplayAlignAddin}
\title{Insert LaTeX displaymath align block}
\usage{
insertDisplayAlignAddin()
}
\description{
Insert LaTeX displaymath blosk with an align environment.
}
