% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertOneColumnAddin.R
\name{insertOneColumnAddin}
\alias{insertOneColumnAddin}
\title{Insert Rmarkdown div for an extra column}
\usage{
insertOneColumnAddin()
}
\description{
Insert Rmarkdown div for a single column into a multi-column layout.
Uses the pandoc \code{:::} ... \code{:::} syntax for divs, with a
\code{{.column}} tag.
}
